feat: 결과 화면 및 러닝 흐름 개선 - API 연결 및 UI 수정
