+++ src/app/index.tsx
@@ -1,8 +1,8 @@
 import { Redirect } from "expo-router";
 
 export default function Index() {
-    return <Redirect href="/(tabs)/home" />;
+    // return <Redirect href="/(tabs)/home" />;
     // return <Redirect href="/course/1" />;
-    // return <Redirect href="/run/1/course" />;
-    // return <Redirect href="/result/7" />;
+    return <Redirect href="/run/7/course" />;
+    // return <Redirect href="/result/7/-1" />;
 }
