+++ src/app/result/[runningId]/[ghostRunningId].tsx
@@ -1,4 +1,5 @@
 import { ShareIcon } from "@/assets/svgs/svgs";
+import { getRun, patchCourseName, patchRunName } from "@/src/apis";
 import StyledChart from "@/src/components/chart/StyledChart";
 import CourseLayer from "@/src/components/map/CourseLayer";
 import MapViewWrapper from "@/src/components/map/MapViewWrapper";
@@ -10,229 +11,292 @@ import NameInput from "@/src/components/ui/NameInput";
 import SlideToDualAction from "@/src/components/ui/SlideToDualAction";
 import StatRow from "@/src/components/ui/StatRow";
 import { Typography } from "@/src/components/ui/Typography";
-import { Course } from "@/src/types/course";
-import { calculateCenter } from "@/src/utils/mapUtils";
+import {
+    calculateCenter,
+    calculateZoomLevelFromSize,
+    convertTelemetriesToCourse,
+} from "@/src/utils/mapUtils";
+import { getDate, getFormattedPace, getRunTime } from "@/src/utils/runUtils";
 import { BottomSheetModal } from "@gorhom/bottom-sheet";
+import { useQuery } from "@tanstack/react-query";
 import { useLocalSearchParams, useRouter } from "expo-router";
-import { useRef, useState } from "react";
+import { useMemo, useRef, useState } from "react";
 import { ScrollView, StyleSheet, View } from "react-native";
 import {
     SafeAreaView,
     useSafeAreaInsets,
 } from "react-native-safe-area-context";
-
-const DATA = Array.from({ length: 100 }, (_, i) => ({
-    distance: i * 10,
-    paceLast30: 40 + 30 * Math.random(),
-    altitude: 100 + 10 * Math.random(),
-    time: i * 10,
-}));
+import Toast from "react-native-toast-message";
 
 export default function Result() {
     const { runningId, ghostRunningId } = useLocalSearchParams();
+    const [isCourseModalOpen, setIsCourseModalOpen] = useState(false);
 
     const bottomSheetRef = useRef<BottomSheetModal>(null);
     const handlePresentModalPress = () => {
         bottomSheetRef.current?.present();
+        setIsCourseModalOpen(true);
     };
 
-    const [recordTitle, setRecordTitle] = useState("월요일 아침 러닝");
+    const { data, isLoading, isError } = useQuery({
+        queryKey: ["result", runningId],
+        queryFn: () => getRun(Number(runningId)),
+        enabled: !!runningId,
+    });
+
+    const {
+        data: ghostData,
+        isLoading: ghostIsLoading,
+        isError: ghostIsError,
+    } = useQuery({
+        queryKey: ["result", ghostRunningId],
+        queryFn: () => getRun(Number(ghostRunningId)),
+        enabled: ghostRunningId !== "-1",
+    });
+
+    const [recordTitle, setRecordTitle] = useState(data?.runningName ?? "");
     const [courseName, setCourseName] = useState("");
 
-    //2025.06.24
-    const date = new Date()
-        .toLocaleDateString("ko-KR", {
-            year: "numeric",
-            month: "2-digit",
-            day: "numeric",
-        })
-        .slice(0, 12)
-        .split(". ")
-        .join(".");
     const router = useRouter();
-    const course = {
-        id: 1,
-        name: "월요일 아침 러닝",
-        coordinates: [
-            [126.85, 37.48],
-            [126.86, 37.49],
-        ],
-    } as Course;
 
-    const center = calculateCenter(
-        course.coordinates.map((coordinate) => ({
-            lat: coordinate[1],
-            lng: coordinate[0],
-        }))
+    const center = useMemo(
+        () =>
+            calculateCenter(
+                data?.telemetries?.map((telemetry) => ({
+                    lat: telemetry.lat,
+                    lng: telemetry.lng,
+                })) ?? []
+            ),
+        [data?.telemetries]
     );
 
     const { bottom } = useSafeAreaInsets();
 
+    if (isLoading || ghostIsLoading) {
+        return <></>;
+    }
+
+    if (isError || ghostIsError) {
+        router.replace("/");
+    }
+
     return (
-        <>
-            <SafeAreaView style={styles.container}>
-                <Header titleText={date} />
-                <ScrollView
-                    contentContainerStyle={styles.content}
-                    keyboardShouldPersistTaps="handled"
-                >
-                    <View style={styles.titleContainer}>
-                        <NameInput
-                            defaultValue="월요일 아침 러닝"
-                            placeholder="제목을 입력해주세요"
-                            onChangeText={setRecordTitle}
-                        />
-                        <ShareIcon />
-                    </View>
-                    <View style={styles.mapContainer}>
-                        <MapViewWrapper
-                            controlEnabled={false}
-                            showPuck={false}
-                            center={center}
-                            zoom={14}
-                        >
-                            <CourseLayer
-                                course={course}
-                                isActive={true}
-                                onClickCourse={() => {}}
-                            />
-                        </MapViewWrapper>
-                    </View>
-                    <View
-                        style={{
-                            paddingHorizontal: 17,
-                        }}
+        data && (
+            <>
+                <SafeAreaView style={styles.container}>
+                    <Header titleText={getDate(data.startedAt)} />
+                    <ScrollView
+                        contentContainerStyle={styles.content}
+                        keyboardShouldPersistTaps="handled"
                     >
-                        <StatRow
+                        <View style={styles.titleContainer}>
+                            <NameInput
+                                defaultValue={data.runningName}
+                                placeholder="제목을 입력해주세요"
+                                onChangeText={setRecordTitle}
+                                onBlur={async () => {
+                                    await patchRunName(
+                                        Number(runningId),
+                                        recordTitle,
+                                        1
+                                    );
+                                }}
+                            />
+                            <ShareIcon />
+                        </View>
+                        <View style={styles.mapContainer}>
+                            <MapViewWrapper
+                                controlEnabled={false}
+                                showPuck={false}
+                                center={center}
+                                zoom={calculateZoomLevelFromSize(
+                                    center.size,
+                                    center.latitude
+                                )}
+                            >
+                                <CourseLayer
+                                    course={convertTelemetriesToCourse(
+                                        data.telemetries ?? []
+                                    )}
+                                    isActive={true}
+                                    onClickCourse={() => {}}
+                                />
+                            </MapViewWrapper>
+                        </View>
+                        <View
                             style={{
-                                paddingVertical: 20,
-                                justifyContent: "space-between",
+                                paddingHorizontal: 17,
                             }}
-                            stats={[
-                                {
-                                    value: "1.45",
-                                    unit: "km",
-                                    description: "전체 거리",
-                                },
-                                {
-                                    value: "25:45",
-                                    unit: "",
-                                    description: "시간",
-                                },
-                                {
-                                    value: "150",
-                                    unit: "spm",
-                                    description: "케이던스",
-                                },
-                                {
-                                    value: "90",
-                                    unit: "kcal",
-                                    description: "칼로리",
-                                },
-                            ]}
-                        />
-                        <Divider direction="horizontal" />
-                        <CollapsibleSection
-                            title="페이스"
-                            defaultOpen={true}
-                            alwaysVisibleChildren={
-                                <StatRow
-                                    style={{
-                                        gap: 20,
-                                    }}
-                                    stats={[
-                                        {
-                                            value: "8'23''",
-                                            unit: "",
-                                            description: "평균",
-                                        },
-                                        {
-                                            value: "10'23''",
-                                            unit: "",
-                                            description: "최고",
-                                        },
-                                        {
-                                            value: "10'23''",
-                                            unit: "",
-                                            description: "최저",
-                                        },
-                                    ]}
-                                />
-                            }
                         >
-                            <StyledChart
-                                data={DATA}
-                                xKey="distance"
-                                yKeys={["paceLast30"]}
+                            <StatRow
+                                style={{
+                                    paddingVertical: 20,
+                                    justifyContent: "space-between",
+                                }}
+                                stats={[
+                                    {
+                                        value: (
+                                            data.recordInfo.distance / 1000
+                                        ).toFixed(2),
+                                        unit: "km",
+                                        description: "전체 거리",
+                                    },
+                                    {
+                                        value: getRunTime(
+                                            data.recordInfo.duration,
+                                            "MM:SS"
+                                        ),
+                                        unit: "",
+                                        description: "시간",
+                                    },
+                                    {
+                                        value: data.recordInfo.cadence.toString(),
+                                        unit: "spm",
+                                        description: "케이던스",
+                                    },
+                                    {
+                                        value: data.recordInfo.calories.toString(),
+                                        unit: "kcal",
+                                        description: "칼로리",
+                                    },
+                                ]}
                             />
-                        </CollapsibleSection>
-                        <Divider direction="horizontal" />
-                        <CollapsibleSection
-                            title="고도"
-                            defaultOpen={true}
-                            alwaysVisibleChildren={
-                                <StatRow
-                                    style={{
-                                        gap: 20,
-                                    }}
-                                    stats={[
-                                        {
-                                            value: "17",
-                                            unit: "m",
-                                            description: "평균",
-                                        },
-                                        {
-                                            value: "+18",
-                                            unit: "m",
-                                            description: "상승",
-                                        },
-                                        {
-                                            value: "-13",
-                                            unit: "m",
-                                            description: "하강",
-                                        },
-                                    ]}
+                            <Divider direction="horizontal" />
+                            <CollapsibleSection
+                                title="페이스"
+                                defaultOpen={true}
+                                alwaysVisibleChildren={
+                                    <StatRow
+                                        style={{
+                                            gap: 20,
+                                        }}
+                                        stats={[
+                                            {
+                                                value: getFormattedPace(
+                                                    data.recordInfo.averagePace
+                                                ),
+                                                unit: "",
+                                                description: "평균",
+                                            },
+                                            {
+                                                value: getFormattedPace(
+                                                    data.recordInfo.highestPace
+                                                ),
+                                                unit: "",
+                                                description: "최고",
+                                            },
+                                            {
+                                                value: getFormattedPace(
+                                                    data.recordInfo.lowestPace
+                                                ),
+                                                unit: "",
+                                                description: "최저",
+                                            },
+                                        ]}
+                                    />
+                                }
+                            >
+                                <StyledChart
+                                    data={data.telemetries}
+                                    xKey="dist"
+                                    yKeys={["pace"]}
                                 />
+                            </CollapsibleSection>
+                            <Divider direction="horizontal" />
+                            <CollapsibleSection
+                                title="고도"
+                                defaultOpen={true}
+                                alwaysVisibleChildren={
+                                    <StatRow
+                                        style={{
+                                            gap: 20,
+                                        }}
+                                        stats={[
+                                            {
+                                                value: data.recordInfo.totalElevation.toString(),
+                                                unit: "m",
+                                                description: "평균",
+                                            },
+                                            {
+                                                value: data.recordInfo.elevationGain.toString(),
+                                                unit: "m",
+                                                description: "상승",
+                                            },
+                                            {
+                                                value: data.recordInfo.elevationLoss.toString(),
+                                                unit: "m",
+                                                description: "하강",
+                                            },
+                                        ]}
+                                    />
+                                }
+                            >
+                                <StyledChart
+                                    data={data.telemetries}
+                                    xKey="dist"
+                                    yKeys={["alt"]}
+                                />
+                            </CollapsibleSection>
+                            <Divider direction="horizontal" />
+                        </View>
+                    </ScrollView>
+                    <SlideToDualAction
+                        onSlideLeft={() => {
+                            router.replace("/");
+                        }}
+                        onSlideRight={async () => {
+                            if (isCourseModalOpen) {
+                                if (data.courseInfo === null) {
+                                    console.log("NO COURSE");
+                                    Toast.show({
+                                        type: "info",
+                                        text1: "코스 정보를 찾을 수 없습니다",
+                                        position: "bottom",
+                                    });
+                                    router.replace("/");
+                                    return;
+                                }
+                                await patchCourseName(
+                                    data.courseInfo.id,
+                                    courseName,
+                                    true
+                                );
+                                // TODO: 코스 등록 후 마이페이지 내 기록으로 이동
+                                router.replace("/");
+                                Toast.show({
+                                    type: "success",
+                                    text1: "코스가 등록 되었습니다",
+                                    position: "bottom",
+                                });
+                            } else {
+                                handlePresentModalPress();
                             }
-                        >
-                            <StyledChart
-                                data={DATA}
-                                xKey="distance"
-                                yKeys={["altitude"]}
-                            />
-                        </CollapsibleSection>
-                        <Divider direction="horizontal" />
-                    </View>
-                </ScrollView>
-                <SlideToDualAction
-                    onSlideLeft={() => {
-                        router.replace("/");
-                    }}
-                    onSlideRight={() => {
-                        console.log(courseName);
-                        handlePresentModalPress();
-                    }}
-                    leftLabel="메인으로"
-                    rightLabel="코스 등록"
-                />
-            </SafeAreaView>
-            <BottomModal
-                bottomSheetRef={bottomSheetRef}
-                bottomInset={bottom + 56}
-                canClose={true}
-                handleStyle={styles.handle}
-            >
-                <View style={styles.bottomSheetContent}>
-                    <NameInput
-                        placeholder="코스명을 입력해주세요"
-                        onChangeText={setCourseName}
+                        }}
+                        leftLabel="메인으로"
+                        rightLabel="코스 등록"
                     />
-                    <Typography variant="body2" color="gray40">
-                        코스를 한 번 등록하면 삭제 및 수정이 어렵습니다
-                    </Typography>
-                </View>
-            </BottomModal>
-        </>
+                </SafeAreaView>
+                <BottomModal
+                    bottomSheetRef={bottomSheetRef}
+                    bottomInset={bottom + 56}
+                    canClose={true}
+                    handleStyle={styles.handle}
+                    onDismiss={() => {
+                        setIsCourseModalOpen(false);
+                    }}
+                >
+                    <View style={styles.bottomSheetContent}>
+                        <NameInput
+                            placeholder="코스명을 입력해주세요"
+                            onChangeText={setCourseName}
+                        />
+                        <Typography variant="body2" color="gray40">
+                            코스를 한 번 등록하면 삭제 및 수정이 어렵습니다
+                        </Typography>
+                    </View>
+                </BottomModal>
+            </>
+        )
     );
 }
